# apps/server/openapi.yaml
openapi: 3.0.3
info:
  title: DiPeO Backend API
  description: |
    DiPeO (Diagrammed People & Organizations) is a visual programming environment
    that executes agent diagrams through a unified backend execution engine.
    
    This API provides endpoints for:
    - Diagram execution with real-time SSE streaming
    - API key management for LLM providers
    - File operations and uploads
    - Conversation memory management
    - Real-time execution monitoring
    
    All diagram nodes execute server-side through the V2 API with consistent
    execution, centralized security, and SSE streaming for real-time updates.
  version: 2.0.0
  license:
    name: MIT
servers:
  - url: http://localhost:8000
    description: Local development server
  - url: https://api.dipeo.io
    description: Production server (example)
...
paths:
  /api/diagrams/execute:
    post:
      tags:
        - diagram
      summary: Execute a diagram
      description: |
        Execute a diagram using the unified backend execution engine with SSE streaming.
        This V2 endpoint consolidates all node execution on the backend.
      operationId: executeDiagram
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - nodes
                - arrows
              properties:
                nodes:
                  type: array
                  items:
                    $ref: '#/components/schemas/DiagramNode'
                  description: List of nodes in the diagram
                arrows:
                  type: array
                  items:
                    $ref: '#/components/schemas/DiagramArrow'
                  description: List of arrows connecting nodes
                persons:
                  type: array
                  items:
                    $ref: '#/components/schemas/Person'
                  description: Person configurations for PersonJob nodes
                apiKeys:
                  type: array
                  items:
                    type: object
                    properties:
                      id:
                        type: string
                      key:
                        type: string
                  description: API keys used in the diagram
                options:
                  type: object
                  properties:
                    continueOnError:
                      type: boolean
                      default: false
                      description: Continue execution even if a node fails
                    allowPartial:
                      type: boolean
                      default: false
                      description: Allow partial execution with unmet dependencies
                    debugMode:
                      type: boolean
                      default: false
                      description: Enable debug mode with detailed logging
      responses:
        '200':
          description: Execution stream started
          content:
            text/event-stream:
              schema:
                type: string
                description: |
                  Server-Sent Events stream with execution updates.
                  Event types include:
                  - execution_started
                  - node_start
                  - node_progress
                  - node_complete
                  - node_skipped
                  - node_error
                  - execution_complete
                  - execution_error
              examples:
                executionStarted:
                  value: |
                    data: {"type": "execution_started", "execution_id": "exec_1234567890", "total_nodes": 5}
                nodeComplete:
                  value: |
                    data: {"type": "node_complete", "nodeId": "node_123", "output": {"result": "Hello"}, "cost": 0.02}
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /api/diagrams/save:
    post:
      tags:
        - diagram
      summary: Save a diagram
      description: Save diagram to the diagrams directory
      operationId: saveDiagram
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - diagram
                - filename
                - format
              properties:
                diagram:
                  type: object
                  description: The diagram object to save
                filename:
                  type: string
                  description: Filename for the saved diagram
                  example: "my-workflow.json"
                format:
                  type: string
                  enum: ["json", "yaml"]
                  description: Format to save the diagram in
      responses:
        '200':
          description: Diagram saved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Diagram saved to files/diagrams/my-workflow.json"
        '500':
          $ref: '#/components/responses/InternalError'
  /api/diagrams/convert:
    post:
      tags:
        - diagram
      summary: Convert diagram format
      description: Convert diagram between different formats (JSON, YAML, LLM-YAML, UML)
      operationId: convertDiagram
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - content
                - from_format
                - to_format
              properties:
                content:
                  type: string
                  description: The diagram content to convert
                from_format:
                  type: string
                  enum: ["yaml", "json", "llm-yaml", "uml"]
                  description: Source format
                to_format:
                  type: string
                  enum: ["yaml", "json", "llm-yaml", "uml"]
                  description: Target format
      responses:
        '200':
          description: Conversion successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  output:
                    type: string
                    description: Converted diagram content
                  message:
                    type: string
        '400':
          $ref: '#/components/responses/ValidationError'
...
components:
  schemas:
    DiagramNode:
      type: object
      required:
        - id
        - type
      properties:
        id:
          type: string
          description: Unique node identifier
          example: "node_123"
        type:
          type: string
          description: Node type (frontend format with "Node" suffix)
          enum: ["startNode", "personJobNode", "personBatchJobNode", "conditionNode", "dbNode", "jobNode", "endpointNode"]
        data:
          type: object
          description: Node properties/configuration
          additionalProperties: true
        position:
          type: object
          properties:
            x:
              type: number
            y:
              type: number
    DiagramArrow:
      type: object
      required:
        - id
        - source
        - target
      properties:
        id:
          type: string
          description: Unique arrow identifier
        source:
          type: string
          description: Source node ID
        target:
          type: string
          description: Target node ID
        sourceHandle:
          type: string
          description: Source connection point identifier
        targetHandle:
          type: string
          description: Target connection point identifier
        label:
          type: string
          description: Arrow label (becomes variable name)
        data:
          type: object
          additionalProperties: true
    Person:
      type: object
      required:
        - id
        - apiKeyId
      properties:
        id:
          type: string
        name:
          type: string
        service:
          type: string
          enum: ["openai", "claude", "gemini", "grok"]
        apiKeyId:
          type: string
        modelName:
          type: string
        systemPrompt:
          type: string
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          description: Error message
        details:
          type: object
          additionalProperties: true
          description: Additional error details
...
  responses:
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: "Validation failed"
            details:
              field: "service"
              reason: "Invalid service name"
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: "Internal server error"
            details:
              message: "An unexpected error occurred"
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT authentication (planned for future release)